feature[canvas]: Write and read from a canvas
Pixels can be written into a canvas and read back out.
Canvas pixels default to black in color.
Current, no out of bounds checks are done, but if a canvas of
w=2 & h=4 is created, the caller of the functions must be aware and
not use a values for w or h that are less than zero, or greater
then either w-1 or h-1.

This is not ideal and I will likely fix this in some manner if it
becomes an issue for me.

The reason I don't do it right now, is the annoyance about calling
macros in when guard clauses in functions.