feature[matrix]: Add submatrices of 4x4 and 3x3 matrices
This probably deserves some more exhaustive testing.
Laying out the submatrices was tedious and had one hard to
spot error that prevented one of these tests from passing for a
while.  There could be more.

It might be worth trying to generate the tests to check all variations.

Maybe that is property testing?  Maybe that would be fun?