feature[clock]: Adds mix task to generate a clock face
This was the reason I started over.  I thought it was a bug in
my code somewhere, but in reality it was a misunderstanding of what
the author meant by "and then move them to the center of your canvas by
adding the coordinates of the center point".

Once I interpreted that correctly, things started to come together.