feature[intersection]: Identify hits of an object
An object is 'hit' by the intersection with the lowest non-negative
t value. The hit is never behind the ray's origin (the camera), so
only positive t values are used.  The lowest t value is the only
one that is 'visible', so that's why it is the result of a hit.